+++ src/components/Good/index.js
@@ -1,0 +1,1 @@
+export * from './Good';
