+++ src/components/SelectButton/SelectButton.jsx
@@ -1,0 +1,10 @@
+export const SelectButton = ({ good, isSelected, selectGood }) => (
+  <button
+    data-cy={isSelected ? 'RemoveButton' : 'AddButton'}
+    type="button"
+    className={`button ${isSelected ? 'is-info' : ''}`}
+    onClick={() => selectGood(isSelected ? null : good)}
+  >
+    {isSelected ? '-' : '+'}
+  </button>
+);
